requirements.txt
Flask==3.0.2
PyGithub==2.1.1
feedparser==6.0.11
python-dotenv==1.0.1
markdown==3.5.2
requests==2.31.0 